### 2025-09-30 #9116 @editorialbot
[PRE REVIEW]: BayCauRETM: R package for Bayesian Causal Inference for Recurrent Event Outcomes
**Submitting author:** @LnnnnYW (<a href="http://orcid.org/0009-0003-8345-9318">Yuqin</a>)
**Repository:** https://github.com/LnnnnYW/BayCauRETM
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
