### 2025-10-01 5fcbe37
revised paper.md
