+++ paper/paper.md
@@ -28,15 +28,15 @@ affiliations:
 
 # Summary
 
-Observational studies are often conducted to estimate the effect of different medical treatments on the rate of a recurrent event outcome within a specified follow-up window. Recurrent events are outcomes that can occur multiple times (e.g. multiple hospitalization events, relapses, etc.) the follow-up window. Causal analysis of recurrent event processes is complicated for several reasons: 1) they are typically observed jointly with a terminal event process such as death. This is strictly dependent on the recurrent event process as, for example, the occurrence of death precludes subsequent recurrence of the outcome of interest. 2) Both the event counts and terminal event process are unobserved if a subject drops out of a study before the end of follow-up - a phenomenon known as censoring. 3) Patients may initiate a treatment at different times during the follow-up, meaning that for each treatment there are as many possible strategies as there are possible initiation times. Finally, 4) since treatments are not randomized in observational studies, formal causal methods such as g-computation are needed to adjust for observed confounders recorded in the data.
+Observational studies often estimate the effect of medical treatments on the rate of a recurrent event within a specific follow-up window. Recurrent events (e.g., repeated hospitalizations, relapses) may occur multiple times during follow-up. Causal analysis is challenging because: 1) Recurrent events are jointly observed with a terminal event (e.g., death), which truncates future recurrences. 2) Both event counts and the terminal process are unobserved under dropout/censoring. 3) Patients initiate treatment at different times, yielding as many strategies as initiation times. Finally, 4) Treatments are not randomized, so formal causal methods are required to adjust for observed confounders.
 
-This paper presents `BayCauRETM`, an `R` package for estimating causal effects of different treatment initiation strategies on a recurrent event outcome in the presence of death and censoring. The user specifies a given treatment initiation time and supplies a `data.frame` with confounders to adjust for, as well as columns for censoring, death, and recurrent event counts. The user then specifies a model for the terminal and recurrent event process using standard `R` regression syntax. With these user inputs, `BayCauRETM` can run a causal adjustment procedure and output adjust expected event rates within a follow-up under the specified treatment initiation time. `BayCauRETM` also provides functions for model diagnostics and visualization.
+This paper presents `BayCauRETM`, an `R` package for estimating causal effects of different treatment initiation strategies on a recurrent event outcome in the presence of death and censoring. Users specify an initiation time and supply a `data.frame` containing confounders and columns for censoring, death, and interval-specific recurrent counts, then define terminal and recurrent models via standard `R` formula syntax. Given these inputs, `BayCauRETM` performs causal adjustment and outputs adjusted expected event rates over follow-up under the specified initiation time. The package also provides diagnostic and visualization utilities. 
 
 Intended users include statisticians, epidemiologists, and health-services researchers analyzing observational data.
 
 # Statement of need
 
[…]
