### 2025-10-01 #9116 @LnnnnYW
I’ve pushed a revised paper.md and checked the two “SKIP DOIs” — there are indeed no DOIs available for those entries.
Larry Dong and Jack Wolf would be our suggestions for the reviewers.
