### 2026-02-01 #9458 @LnnnnYW
Hello @larryshamalama and @jackmwolf ,

Thank you both very much for your careful and thoughtful comments. We truly appreciate the time you put into reviewing our work. We have now made revisions addressing all of your comments, and the updated version is ready for re-check and review at your convenience.

Hi @larryshamalama ,
Below we respond to your comments point by point:
1. We have added a `Quickstart` section to both the README and the paper. This example can be run directly and clearly demonstrates the overall pipeline of the package.
2. We expanded the Statement of Need to include a discussion of existing related software, with additional details on available methods and packages that address similar problems.
3. We incorporated your pull request and carefully revised the manuscript for clarity, organization, and writing.
4. The community guidelines section has now been added to the end of the paper.

And, Hi @jackmwolf ,
We have responded to your detailed initial review in the [associated issue](https://github.com/LnnnnYW/BayCauRETM/issues/1) and addressed all of the points raised there. We hope the current version fully resolves your comments.

Thank you again to both of you for the constructive and insightful feedback. Please feel free to let us know if any issues remain or if further clarification would be helpful. We would be very happy to make additional improvements if needed.
