### 2026-02-18 5a71810
Update g_computation.Rd
