### 2026-03-22 #9458 @larryshamalama
> > > > Hi [@adithirgis](https://github.com/adithirgis), I edited my checklist to push this towards acceptance. I rarely check my GitHub notifications these days, so please email me at [larry.dong@mail.utoronto.ca](mailto:larry.dong@mail.utoronto.ca) if anything else is required from my end
> > >
> > >
> > > Hi [@larryshamalama](https://github.com/larryshamalama) thank you for sharing your email ID. I see that there are 2 unchecked items in the checklist (**_State of the field_** and **_Quality of writing_**). Would you please update them. Thanks!
> >
> >
> > Thank you all for your careful review! [@adithirgis](https://github.com/adithirgis), perhaps this can be communicated via e-mail since Larry indicated he doesn't check GitHub?
>
> Hi [@stablemarkets](https://github.com/stablemarkets) I did email the same to [@larryshamalama](https://github.com/larryshamalama). I will send a reminder in a couple of days :) Thank you!

Done! Apologies those were missed
