### 2026-03-23 #9458 @LnnnnYW
Hi @adithirgis, Thank you! I have completed the requested final steps.

- All reviewer comments have been addressed and resolved.
- The title has been checked and contains no special formatting.
- A GitHub release (v1.0.1) was created.
- The release has been archived in Zenodo with correct metadata matching the paper title and author list.

DOI: https://doi.org/10.5281/zenodo.19177819
Version tag: v1.0.1
