+++ paper/paper.bib
@@ -17,7 +17,7 @@ journal = {Journal of the Royal Statistical Society: Series C (Applied Statistic
 volume = {55},
 number = {1},
 pages = {31-39},
-doi = {https://doi.org/10.1111/j.1467-9876.2005.00529.x},
+doi = {10.1111/j.1467-9876.2005.00529.x},
 year = {2006}
 }
 
@@ -28,7 +28,7 @@ year = {2006}
   volume={38},
   pages={1--28},
   year={2011},
-  doi = {https://doi.org/10.18637/jss.v038.i08}
+  doi = {10.18637/jss.v038.i08}
 }
 
 @article{Chiou2023,
@@ -51,7 +51,7 @@ year = {2006}
   number={2},
   pages={91--115},
   year={2002},
-  doi = {https://doi.org/10.1191/0962280202SM276ra}
+  doi = {10.1191/0962280202SM276ra}
 }
 
 
@@ -91,7 +91,7 @@ author = {Lyu, Tianmeng and Bornkamp, Björn and Mueller-Velten, Guenther and Sc
 title = {Bayesian inference for a principal stratum estimand on recurrent events truncated by death},
 journal = {Biometrics},
 year = 2022,
-doi = {https://doi.org/10.1111/biom.13831},
+doi = {10.1111/biom.13831},
 }
 
 @article{janvin2024,
