### 2026-03-26 #9458 @LnnnnYW
Thanks @adithirgis , I've checked the DOIs and they seem to be good now!
