### 2026-04-01 #9458 @adithirgis
> Hi [@LnnnnYW](https://github.com/LnnnnYW) for the Zenodo I see this as the affiliation.
>
> <img alt="Image" width="647" height="149" src="https://private-user-images.githubusercontent.com/63432077/572267934-a7fbab5a-f404-4386-a767-d076e5a5adeb.png?jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NzUwMTUzMjYsIm5iZiI6MTc3NTAxNTAyNiwicGF0aCI6Ii82MzQzMjA3Ny81NzIyNjc5MzQtYTdmYmFiNWEtZjQwNC00Mzg2LWE3NjctZDA3NmU1YTVhZGViLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNjA0MDElMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjYwNDAxVDAzNDM0NlomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTYwYWNhMDQyNzI0NjA3YTNkZjY1MWJjMWFhMmY1YjMzNGFjNWM1Njc0YTRkZDRmYzNiNDZkMTNkYTM0ZmE4YjcmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.DQkVYR14e-vU2WnVtDBU3onlSH19oAEi62GgI3sqFQA">

While in the paper it is as shown below. I wanted to request you to match the Zenodo affiliation to exact same - Department of Biostatistics....
