### 2026-04-20 #9458 @crvernon
@LnnnnYW the submission is looking great.  I just need you to address the following before I accept for publication:

In your paper:
- [ ] LINE 145:  The "r" as in the R programming language should be capitalized. You can do this in your bib file by wrapping the text you want to maintain the formatting of in curly brackets.
- [ ] LINE 152:  The "r" as in the R programming language should be capitalized.

No need to create another release or mint a new archive for these changes.

Let me know when they are done and I'll accept after confirmation.

Best,

Chris
