### 2026-04-20 3a839d4
Update paper.bib
Capitalize R in reference
