+++ paper/paper.bib
@@ -22,7 +22,7 @@ year = {2006}
 }
 
 @article{jackson2011,
-  title={Multi-state models for panel data: the msm package for R},
+  title={Multi-state models for panel data: the msm package for {R}},
   author={Jackson, Christopher},
   journal={Journal of statistical software},
   volume={38},
@@ -32,7 +32,7 @@ year = {2006}
 }
 
 @article{Chiou2023,
- title={Regression Modeling for Recurrent Events Possibly with an Informative Terminal Event Using R Package reReg},
+ title={Regression Modeling for Recurrent Events Possibly with an Informative Terminal Event Using {R} Package reReg},
  volume={105},
  doi={10.18637/jss.v105.i05},
   number={5},
